Update _creation_functions.py
Float values passes as an argument in def arange on line 75 to ensure that multiplying numpy.arange with a large number does effective scaling in windows which does not happen otherwise

+++ numpy/array_api/_creation_functions.py
@@ -73,10 +73,10 @@ def asarray(
 
 
 def arange(
-    start: Union[int, float],
+    start: Union[float, float],
     /,
-    stop: Optional[Union[int, float]] = None,
-    step: Union[int, float] = 1,
+    stop: Optional[Union[float, float]] = None,
+    step: Union[float, float] = 1,
     *,
     dtype: Optional[Dtype] = None,
     device: Optional[Device] = None,
